feat(client): reorganize, introduce test, stub a client module
