+++ lib/killer.js
