+++ test/client.it.js
@@ -1,0 +1,8 @@
+const { expect } = require('chai');
+const tscWatchClient = require('../client');
+
+describe('Client', () => {
+  it('Should be consumable as a module', () => {
+    expect(tscWatchClient).to.exist;
+  });
+});
